Crée un fichier particulier pour la méthode
